consolidate unit test
