--- content_md_test.go
@@ -1,88 +1,0 @@
-package errz
-
-import (
-	"strings"
-	"sync"
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-)
-
-// TitleCacheReset clears the title cache (for test use only)
-func TitleCacheReset() {
-	titleCache = sync.Map{}
-}
-
-func TestGenerateMarkdownContent_ValidInput(t *testing.T) {
-	TitleCacheReset()
-	defer TitleCacheReset()
-
-	errorsMap := map[string]Error{
-		"ERR001": {
-			Code:  "ERR001",
-			Msg:   "Invalid input | bad format",
-			Cause: "Input contained unexpected value",
-		},
-		"ERR002": {
-			Code:  "ERR002",
-			Msg:   "Timeout `network`",
-			Cause: "Service did not respond",
-		},
-	}
-
-	md, err := generateMarkdownContent("core-api", errorsMap)
-	assert.NoError(t, err)
-	assert.Contains(t, md, "# Core-Api Errors")
-	assert.Contains(t, md, "| ERR001 | Invalid input \\| bad format |")
-	assert.Contains(t, md, "- **Cause**: Input contained unexpected value")
-	assert.Contains(t, md, "## ERR002")
-	assert.Contains(t, md, "Timeout \\`network\\`")
-}
-
-func TestGenerateMarkdownContent_InvalidDomain(t *testing.T) {
-	TitleCacheReset()
-	defer TitleCacheReset()
-
-	_, err := generateMarkdownContent("bad domain", map[string]Error{})
-	assert.ErrorIs(t, err, errInvalidDomainName)
-
-	_, err = generateMarkdownContent(" ", map[string]Error{})
-	assert.ErrorIs(t, err, errInvalidDomainName)
-}
-
-func TestGenerateMarkdownContent_EmptyErrors(t *testing.T) {
-	TitleCacheReset()
-	defer TitleCacheReset()
-
-	md, err := generateMarkdownContent("example", map[string]Error{})
-	assert.Error(t, err)
-	assert.Empty(t, md)
-	assert.EqualError(t, err, "no error definitions provided for markdown generation")
-}
-
-func TestGenerateMarkdownContent_Sorting(t *testing.T) {
-	TitleCacheReset()
-	defer TitleCacheReset()
-
-	errorsMap := map[string]Error{
-		"B": {Code: "B"},
-		"A": {Code: "A"},
-	}
-
-	md, err := generateMarkdownContent("domain", errorsMap)
-	assert.NoError(t, err)
-	firstIdx := strings.Index(md, "## A")
-	secondIdx := strings.Index(md, "## B")
-	assert.Less(t, firstIdx, secondIdx)
-}
-
-func TestNormalizeMarkdownTitle_CachesCorrectly(t *testing.T) {
-	TitleCacheReset()
-	defer TitleCacheReset()
-
-	title1 := normalizeMarkdownTitle("my-service")
-	title2 := normalizeMarkdownTitle("my-service")
-
-	assert.Equal(t, "# My-Service Errors\n\n", title1)
-	assert.Equal(t, title1, title2) // from cache
-}
